refactor: use separate files for commands and handlers
