+++ src/Transactions/BankingApp.Transactions.API/Features/AccountCreation/CreateAccountCommand.cs
@@ -1,0 +1,5 @@
+﻿using MediatR;
+
+namespace BankingApp.Transactions.API.Features.AccountCreation;
+
+public record CreateAccountCommand(Guid HolderId, string Name, string Document, string Token, string Currency) : IRequest;
