+++ src/Transactions/BankingApp.Transactions.API/Features/AccountUpdated/UpdateAccountCommand.cs
@@ -1,0 +1,5 @@
+﻿using MediatR;
+
+namespace BankingApp.Transactions.API.Features.AccountUpdated;
+
+public record UpdateAccountCommand(Guid HolderId, string? Name, string? Document, string? Token, string? Currency) : IRequest;
