+++ src/Transactions/BankingApp.Transactions.API/Features/Withdraws/CreateWithdrawCommandHandler.cs
@@ -1,0 +1,32 @@
+﻿using BankingApp.Transactions.API.Infrastructure;
+using BankingApp.Transactions.Domain.Exceptions;
+using MediatR;
+using Microsoft.EntityFrameworkCore;
+
+namespace BankingApp.Transactions.API.Features.Withdraws;
+
+public class CreateWithdrawCommandHandler : IRequestHandler<CreateWithdrawCommand, WithdrawTransactionResponse>
+{
+    private readonly AccountsDbContext _context;
+
+    public CreateWithdrawCommandHandler(AccountsDbContext context)
+    {
+        _context = context;
+    }
+
+    public async Task<WithdrawTransactionResponse> Handle(CreateWithdrawCommand request, CancellationToken cancellationToken)
+    {
+        var account = await _context.Accounts
+            .FirstOrDefaultAsync(account => account.Holder.Token == request.Token, cancellationToken)
+            .ConfigureAwait(continueOnCapturedContext: false);
+
+        if (account is null)
+        {
+            throw new AccountNotFoundException($"Account not found for token {request.Token}");
+        }
+
+        var transaction = account.Withdraw(request.Amount, DateTime.UtcNow);
+
+        return new WithdrawTransactionResponse(transaction.Id, transaction.Type.Value);
+    }
+}
